# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Nose IoT System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RespondentGender =
  (typeof RespondentGender)[keyof typeof RespondentGender];

export const RespondentGender = {
  male: "male",
  female: "female",
} as const;

export type RespondentStatus =
  (typeof RespondentStatus)[keyof typeof RespondentStatus];

export const RespondentStatus = {
  diabetes: "diabetes",
  "non-diabetes": "non-diabetes",
} as const;

export interface Respondent {
  id: number;
  name: string;
  age: number;
  gender: RespondentGender;
  status: RespondentStatus;
  createdAt: string;
}

export type CreateRespondentInputGender =
  (typeof CreateRespondentInputGender)[keyof typeof CreateRespondentInputGender];

export const CreateRespondentInputGender = {
  male: "male",
  female: "female",
} as const;

export type CreateRespondentInputStatus =
  (typeof CreateRespondentInputStatus)[keyof typeof CreateRespondentInputStatus];

export const CreateRespondentInputStatus = {
  diabetes: "diabetes",
  "non-diabetes": "non-diabetes",
} as const;

export interface CreateRespondentInput {
  name: string;
  age: number;
  gender: CreateRespondentInputGender;
  status: CreateRespondentInputStatus;
}

export type SessionState = (typeof SessionState)[keyof typeof SessionState];

export const SessionState = {
  pending: "pending",
  purging: "purging",
  ready: "ready",
  sampling: "sampling",
  completed: "completed",
  error: "error",
} as const;

export interface Session {
  id: number;
  respondentId: number;
  sessionNumber: number;
  state: SessionState;
  mq3AlcoholPpm?: number | null;
  mq4MethanePpm?: number | null;
  mq138AcetonePpm?: number | null;
  tgs2602Voc?: number | null;
  samplesCount?: number | null;
  createdAt: string;
  completedAt?: string | null;
}

export type SessionWithRespondent = Session & {
  respondent: Respondent;
};

export type RespondentWithSessions = Respondent & {
  sessions: Session[];
};

export interface CreateSessionInput {
  respondentId: number;
}

export interface SensorDataInput {
  mq3AlcoholPpm: number;
  mq4MethanePpm: number;
  mq138AcetonePpm: number;
  tgs2602Voc: number;
  samplesCount: number;
}

export type DeviceCommandCommand =
  (typeof DeviceCommandCommand)[keyof typeof DeviceCommandCommand];

export const DeviceCommandCommand = {
  idle: "idle",
  start_session: "start_session",
  purge: "purge",
  sample: "sample",
  stop: "stop",
} as const;

export interface DeviceCommand {
  command: DeviceCommandCommand;
  sessionId?: number | null;
  purgeDuration?: number | null;
  sampleDuration?: number | null;
}

export type DeviceStatusInputState =
  (typeof DeviceStatusInputState)[keyof typeof DeviceStatusInputState];

export const DeviceStatusInputState = {
  idle: "idle",
  purging: "purging",
  ready: "ready",
  sampling: "sampling",
  completed: "completed",
  error: "error",
} as const;

export interface DeviceStatusInput {
  state: DeviceStatusInputState;
  sessionId?: number | null;
  message?: string | null;
}

export type DeviceStatusState =
  (typeof DeviceStatusState)[keyof typeof DeviceStatusState];

export const DeviceStatusState = {
  idle: "idle",
  purging: "purging",
  ready: "ready",
  sampling: "sampling",
  completed: "completed",
  error: "error",
} as const;

export interface DeviceStatus {
  state: DeviceStatusState;
  sessionId?: number | null;
  message?: string | null;
  lastSeen?: string | null;
  isOnline: boolean;
}