# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: E-Nose IoT System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /respondents:
    get:
      operationId: listRespondents
      tags: [respondents]
      summary: List all respondents
      responses:
        "200":
          description: List of respondents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Respondent"
    post:
      operationId: createRespondent
      tags: [respondents]
      summary: Create a new respondent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRespondentInput"
      responses:
        "201":
          description: Respondent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Respondent"
  /respondents/{id}:
    get:
      operationId: getRespondent
      tags: [respondents]
      summary: Get a respondent by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Respondent found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RespondentWithSessions"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Respondent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        gender:
          type: string
          enum: [male, female]
        status:
          type: string
          enum: [diabetes, non-diabetes]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - age
        - gender
        - status
        - createdAt
    CreateRespondentInput:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        gender:
          type: string
          enum: [male, female]
        status:
          type: string
          enum: [diabetes, non-diabetes]
      required:
        - name
        - age
        - gender
        - status
    Session:
      type: object
      properties:
        id:
          type: integer
        respondentId:
          type: integer
        sessionNumber:
          type: integer
        state:
          type: string
          enum: [pending, purging, ready, sampling, completed, error]
        mq3AlcoholPpm:
          type: number
          nullable: true
        mq4MethanePpm:
          type: number
          nullable: true
        mq138AcetonePpm:
          type: number
          nullable: true
        tgs2602Voc:
          type: number
          nullable: true
        samplesCount:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - respondentId
        - sessionNumber
        - state
        - createdAt
...
    RespondentWithSessions:
      allOf:
        - $ref: "#/components/schemas/Respondent"
        - type: object
          properties:
            sessions:
              type: array
              items:
                $ref: "#/components/schemas/Session"
          required:
            - sessions
...