# lib/api-zod/src/generated/types/createRespondentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Nose IoT System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRespondentInputGender } from "./createRespondentInputGender";
import type { CreateRespondentInputStatus } from "./createRespondentInputStatus";

export interface CreateRespondentInput {
  name: string;
  age: number;
  gender: CreateRespondentInputGender;
  status: CreateRespondentInputStatus;
}