# lib/api-zod/src/generated/types/respondent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Nose IoT System API
 * OpenAPI spec version: 0.1.0
 */
import type { RespondentGender } from "./respondentGender";
import type { RespondentStatus } from "./respondentStatus";

export interface Respondent {
  id: number;
  name: string;
  age: number;
  gender: RespondentGender;
  status: RespondentStatus;
  createdAt: Date;
}