# lib/api-zod/src/generated/types/respondentWithSessions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Nose IoT System API
 * OpenAPI spec version: 0.1.0
 */
import type { Respondent } from "./respondent";
import type { Session } from "./session";

export type RespondentWithSessions = Respondent & {
  sessions: Session[];
};